Fixes bug in Timing module tests.
Using AutoResetEvent or any other kind of EventWaitHandle is not a great idea because of a race between the EventWaitHandle being set twice and waited upon. Specifically, it's possible for the timer to run twice before the waiting thread observes the unblock operation.  Switching to CountdownEvent instead, which provides much more reliable results.