refactor(ReactWindows): Reverting changes outside ReactWindows folder
Starting to move towards a model where ReactWindows can be layered on top of React Native without requiring any core JS changes. The approach here is to take any of the files that had been modified for the Windows platform and copy them into the ReactWindows sub-folder, applying the .windows.js extension. The packager is already built to register the 'generic' module separately from the 'windows' module.