breaking(ReactStylesDiffMap): removing unnecessary abstraction
The ReactStylesDiffMap is not a useful abstraction. Everything in react-native-windows is tightly coupled with Newtonsoft.Json. If we decide to abandon that dependency, then we should revisit a ReactNative specific JSON data model abstraction. In the meantime, since we're making breaking changes to IViewManager anyway, it's a good time to eliminate this class.