[rn][win] Ensure root views can be properly unloaded
Summary:
This adds `ReactViewHost::PostInQueue` method that *safely* calls the callback as long as the `ReactViewHost` is still alive. It was unsafe before as it would capture the raw `this` pointer, which could result in a crash (I was able to observe this).

However, for *unloading* the view, we actually want to capture the `ReactViewHost` strongly to ensure that it gets fully detached from the `ReactHost` and that `AppRegistry.unmountApplicationComponentAtRootTag()` is called.

Test Plan: Confirmed no more warning (added in D26616622) that logs when `AppRegistry.unmountApplicationComponentAtRootTag()` is not called.

Reviewers: ericroz, jacke, mylando

Reviewed By: ericroz

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D26643176

Signature: 26643176:1614217452:b757796e759fda36e8732650854ea3ab341cf8bf