Fixes case where focus commands may be ignored (#7304)
* Fixes case where focus commands may be ignored

Given JS state is async from native state, it is possible that a control
may lose native focus, but the TextInputState module that manages focus
for TextInput components may not yet have received the event, and as a
result may ignore commands to focus the TextInput. This change removes a
check from TextInputState that would prevent focus commands from
executing.

Fixes #7303

* Change files