[rn][win] Adds support for scroll anchoring in react-native-windows
Summary:
Chat apps typically requires two scroll anchoring features:
1. Scroll anchoring while scrolled away from the bottom edge of the chat window.
2. Bottom edge anchoring when "docked" at the bottom edge of the chat window, so new messages automatically come into view.

This diff achieves feature 1. by always setting `UIElement.CanBeScrollAnchor` on `ScrollViewContent` children. The side-effect of this is that non-anchored behavior will no longer be possible, though I'm not sure this is a problem as that is the default Web behavior.

This diff achieves feature 2. by adding a `nativeInverted` prop for `ScrollView`. When combined with `style={{flexDirection: 'column-reverse'}}` on a FlatList, it achieves effectively the same behavior as the `inverted` prop.

Landing changes to OSS will take some time, so this is a good temporary path.

Test Plan:
https://fb.quip.com/4zzqAVGEOLK1

Sticks to bottom for incoming messages and typing indicator:

{F507395097}

Scrolls to bottom when not at bottom:

{F507398469}

Anchors when loading new content:

{F507399405}


Reviewers: skyle, lyahdav, mylando, mowang

Reviewed By: skyle

Subscribers: taskcreeper, eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D27001566

Tasks: T86030936, T86021290

Signature: 27001566:1615918992:078445634cacf466d5854d996e9d905766ee3d51