[rn][win] Ensure keyDownEvents are handled before submitKeyEvents (fix from react-native-windows#7333)
Summary:
Originally, I thought we could eliminate the need for `keyDownEvents` using `clearTextOnSubmit` and `submitKeyEvents`. However, we still need to conditionally disable submit when the mentions popover is open.

There's a bug in react-native-windows where even when the `keyDownEvents` handled event phase is set to `captured`, the TextInput control will still receive the event. The reason for this is that the `PreviewKeyDown` handler for firing `onSubmitEditing` events is registered *before* the `PreviewKeyDown` handler for `keyDownEvents`.

This change ensures that if `keyDownEvents` are set on TextInput, that the handler for `keyDownEvents` fires before the handler for `onSubmitEditing`.

Test Plan:

1. Clone into FBS
2. Ensure pressing 'Enter' when mentions popover is open closes the popover and does not submit.

{F510418604}


Reviewers: skyle, lyahdav, mylando

Reviewed By: skyle

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D27081141

Tasks: T85628355

Signature: 27081141:1615917267:1448eb2cb4a9dd7a9e7fa1cdae49808319c14b81