Reverting complex behavior in ScrollView anchoring
Summary:
We added the following 3 things to the ScrollView anchoring diff:
1. Logic to turn on and off anchoring when you reached "the top" of the scroll viewer
2. Logic to suppress `onScroll` events for anchoring
3. Logic to emit `onScroll` events when "inverted" offsets change, even when view port did not.

Items 2 and 3 do not seems to cause any issues in inverted scrolling scenarios. They are both "should haves" in the scroll anchoring / inverted doc I posted here:
https://gist.github.com/rozele/47a2ad44f6d44561da1d293454ae8418

Item 1 is a bit different. It adds a lot of complexity for what it actually solves. Absent this change, there is a possibility that the scroll view won't anchor to the bottom when the window is resized (particularly when items inside the scroll view change size). I will be filing a task to fix this bug separately, as the current approach doesn't always work as expected.

Test Plan: https://fb.quip.com/4zzqAVGEOLK1

Reviewers: skyle, lyahdav, mylando

Reviewed By: skyle

Subscribers: taskcreeper, eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D27166248

Signature: 27166248:1616095356:2ae3f4772c2e8817117c4c868aef05cfd18a9a27