[rn][win] Refactors ScrollView anchoring to minimize diff against ScrollViewManager
Summary:
Moves all of the logic to manage view / edge anchoring, scroll restarting, offset flipping, etc. to the ScrollViewViewChanger implementation.

This is a bit of churn as we removed the `SetContentScrollAnchors` functionality to unset view anchoring when scrolled to the bottom edge in D27166248. Re-introducing it as it seems view anchoring sometimes takes precedent over edge anchoring even when we're not resizing the window.

Test Plan: https://fb.quip.com/4zzqAVGEOLK1

Reviewers: skyle, lyahdav, mylando

Reviewed By: skyle

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D27181238

Signature: 27181238:1616179134:a7928775a451342ef6fb89785059afa4803d2ffd