[rn][win] Manual cherry-pick ce85295f: Suppress onLayout event when layout has not changed (#7470)
Summary:
* Suppress onLayout event when layout has not changed

Yoga sometimes sets `hasNewLayout` on nodes even when their layout has
not changed. One common case where this occurs is in lists. Any time the
parent container of a list is dirtied, even for something that does not
affect it's dimensions, all elements in the list go through a layout
pass and set `hasNewLayout`. This change at least reduces the number of
events sent to JS when layout has not really changed, though a more
effective fix would be to have Yoga be a little smarter about when to
layout. If you're interested, here's where Yoga decides to update layout
on all stretch aligned children of a parent container:
https://github.com/facebook/yoga/blob/master/yoga/Yoga.cpp#L3104-L3107

iOS and Android already do this, so this change should serve to make
things more consistent:
iOS:
https://github.com/facebook/react-native/blob/0f4f91766339f2cf4189446d7d575493d33b4009/React/Views/RCTShadowView.m#L305
Android:
https://github.com/facebook/react-native/blob/6e6443afd04a847ef23fb6254a84e48c70b45896/ReactAndroid/src/main/java/com/facebook/react/uimanager/ReactShadowNodeImpl.java#L363-L367

* Updating with results of yarn format

* Change files

* Fixes nit

Test Plan:
Zeratul compiles. There should be a slight benefit to CPU utilization and memory usage as fewer layout events will be queued. This is particularly impactful for vertical resizing of the window and new messages arriving (neither of which result in actual layout changes for most items in the chat thread).

For good measure, here is a grid view that remains responsive:

https://pxl.cl/1Ljhv



Reviewers: skyle, lyahdav, mylando, #archon-rn

Reviewed By: skyle

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D29565630

Signature: 29565630:1625622086:318e6d6231dde99ed8098e894e06bb578d3b81fb