Refines algorithm for hit testing Text
This algorithm implements binary search over the characters in the
TextBlock to find the character whose bounds contain the pointer point.
The parent of that character is the correct target for the pointer
event. The previous algorithm was O(N) in the number of Inline nodes in
the TextBlock sub-tree. This algorith is O(log(C)) in the number of
characters in the TextBlock. In practice this will probably be a bit
more expensive, but solves bugs in the previous algorithm related to hit
testing wrapped text.

Fixes #5151