Fixes crash from tearing down pressable text
Summary:
We recently added functionality to support pressable wrapped text. The approach uses reference counting, so a hit test only occurs on a run (or it's virtual text ancestors) if it has a pressable descendant. When the pressable run is cleaned up, we decrement the ref counts of it's ancestors.

We overlooked the cleanup that occurs when the React Native instance is being torn down (e.g., on app exit). When the app is being torn down, there's a race between the views being cleaned up and the native module being cleaned up. We don't actually care about having a proper ref count in this tear down case, because all nodes in the view tree are being destroyed.

Thus, the fix is just to ensure we have a reference to the native UI manager host before walking up the virtual text tree.

Test Plan:
1. Open a view with pressable text
2. Close the app many times
3. Ensure the exception does not occur

Reviewers: skyle, lyahdav, jacke, mylando

Reviewed By: skyle

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28019173

Tasks: T89364014

Signature: 28019173:1619495211:44f65bc1757154036513fa9bb5a755f6073c00f1