[rn][win] Enable line height less than max font height
Summary:
By default, the XAML TextBlock uses `LineStackingStrategy::MaxHeight`, which means that the lines are stacked according to the maximum height of a character (this likely corresponds to the maximum possible height of a character including ascenders and descenders).

Side note, the font size seems to correspond to the cap height for the font with a bit of extra buffer for "normal" ascender and descenders, so the actual line height when `LineStackingStrategy::MaxHeight` is used is always greater than the font size. For example, the defaults for a XAML TextBlock are font size 14px and Segoe UI for font family. The line height works out to be 18.62px.

React Native allows you to set the `lineHeight` prop, but unfortunately the formula for line height when `LineStackingStrategy::MaxHeight` is used works out to be something like `Math.Max(fontHeight, lineHeight)`, so any value less than the computed maximum font height is "ignored".

This change switches the strategy to use `LineStackingStrategy::BlockLineHeight` when a RN prop value for `lineHeight` is supplied, so the user has explicit control over the line height, even if the value given is less than the computed max font height.

Test Plan:
Measure text height in Typography example of UI Explorer:
|Before|After|
| {F611236287} | {F611235971} |

 Notice that before the change, even though the line height was set to 13px, the actual height of the TextBlock was 29.2617 pixels. After the change, when the line height was set to 13px, the height of the 2-line TextBlock was the expected 26px.



Reviewers: skyle, lyahdav, catfeng, mylando

Reviewed By: skyle

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28028037

Tasks: T89709736

Signature: 28028037:1619535849:55b0a77be8e2e9a596889078bcd296ec9af2f0a3