[RNDesktop][Windows] Ensure selectable text does not fire onMouseEnter/Leave events for its ancestors
Summary:
We subscribe each selectable TextBlock to it's own TouchEventHandler to emit otherwise handled pointer events. The fix overlooked the possibility that the TouchEventHandler for the selectable TextBlock would be able to emit events for it's ancestors.

This change fixes the behavior so the TouchEventHandler subscribed to the TextBlock can only emit onMouseEnter and onMouseLeave events for the TextBlock and it's virtual text descendants.

Test Plan:
1. Clone into fbsource
2. Build Zeratul
3. Confirm the hover behavior for message bubbles works as expected
|Before|After|
|{F611301655}|{F611300715}

Reviewers: skyle, lyahdav, mylando

Reviewed By: mylando

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28039526

Tasks: T89082326

Signature: 28039526:1619555706:c5ea72081ff792cfab42294912003cd6edd7c885