Improves hit test algorithm
This change modifies the algorithm in the following ways:
1. If the entire run is unidirectional on one line, same performance as
previously implemented.
2. If the entire run is LTR (or RTL in FlowDirection::RightToLeft), the
search is O(1) as we check 3 rects (first line, middle lines, last
line).
3. If the run starts or ends with RTL text, we need to find the
character closest and farthest from the origin (respectively) to get the
RTL bounding boxes. This is O(log(c)).

So overall, the algorithm is O(log(c)) in the size of pressable runs
if-and-only-if it starts or ends with RTL text, is multiline, or, if not
single-line, is bidirectional.