[rn][win] Prevent flyout dismissal from stealing focus
Summary: When closing a flyout, react-native-windows attempts to focus the target element of the flyout. This results in the bug referenced in T88704756. This change ensures that the XAML root associated with the flyout target element is still in focus before attempting to re-focus the target.

Test Plan:
|Before|After|
|{F616424978}|{F616424541}|

Reviewers: skyle, lyahdav, mylando

Reviewed By: lyahdav

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28476154

Tasks: T88704756

Signature: 28476154:1621269807:429a9f3d562dbe45ee4e2e999780f950dc3c1696