[rn][win] Adds zoomFromCenter native prop
Summary:
The implementation adds a `zoomFromCenter` prop that, when set to true, ensures the center point of the view port remains fixed while zooming in.

This is a temporary workaround while we don't have a fork of react-native-windows JS code. The ideal solution here would to be either:
1. Implement the `zoomToRect` command for Windows
2. Implement an overload for the `scrollTo` command that also accepts a `zoomScale` parameter

For case 1, there is an invariant in react-native-windows JS that limits the usage of `zoomToRect` to iOS, and for 2 we would need to update the JS for the ScrollView component to support the additional `zoomScale` property in `scrollTo`.


Test Plan:
1. Clone into FBS
2. Update the media viewer to use the native prop
3. Ensure zoom operations keep a fixed center point
|Before|After|
|{F616600015}|{F616607028}|

Reviewers: skyle, lyahdav, mylando, olgarithm

Reviewed By: olgarithm

Subscribers: taskcreeper, eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28497244

Tasks: T91002472

Signature: 28497244:1621378234:3f63c6a9d46a0b128cc2942cf0418e645a35165d