[rn][win] Fix crash related to TextBox context menu
Summary:
In XAML Islands, any context menu with a sub-menu has the potential to
crash due to a known bug:
https://github.com/microsoft/microsoft-ui-xaml/issues/3529

This diff fixes the issue for the proofing AppBarButton in the
TextCommandBarFlyout on TextBox, which has a sub-menu for spell check
suggestions.

Test Plan:
App used to crash readily while moving the mouse around the context menu for TextBox. Now it does not:
|Before|After|
|{F619024504}|{F619024729}|

Reviewers: skyle, lyahdav, mylando

Reviewed By: lyahdav

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28720068

Tasks: T90526985

Signature: 28720068:1622057110:094b9d235a2710f45e613d5d37e7a55efeb53353