[rn][win] Prefer TextBox handling of Paste
Summary:
In D27794270, we added functionality to allow pasting of attachments.
This change ensures that if the TextBox can handle the content on the
clipboard, we allow the TextBox to do so rather than pasting the content
as an attachment.

Test Plan:
Text content pasted as text
|Before|After|
|{F621794690}|{F621793767}|

Images pasted as images and mixed content pasted as text
{F621799025}


Reviewers: skyle, lyahdav, mylando, olgarithm

Reviewed By: skyle

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28901475

Tasks: T91948838

Signature: 28901475:1622824469:fd3067f6b7af6b7caebd2bdf002bc8db336af49e