[rn][win] Allow i18n for custom Paste label and tooltip
Summary:
We previously hardcoded the Paste label and tooltip description for the TextBox context menu when enabling the `onPaste` for attachments in the Zeratul composer. This change adds a few more options to localize the Paste label and tooltip.

The ideal option would be to consume the localized values for paste from XAML. They likely exist in the MUI file for XAML controls at `C:\Windows\System32\{locale}\Windows.UI.Xaml.Controls.dll.mui`, however the string resource indices are not guaranteed to be stable nor are the indices themselves opaque, and the mechanism to these strings is non-trivial to add to react-native-windows. Taking this approach also allows us to customize the tooltip for adding an attachment.

Test Plan:
Ensure custom tooltip message for attachments is shown:
|Before|After|
|{F621828167}|{F621828227}|

We'll need to wait for the fbt translations to land before we can see the effect in different languages.

Reviewers: skyle, lyahdav, mylando

Reviewed By: lyahdav

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D28902886

Tasks: T89055584

Signature: 28902886:1622838611:41a06f939c338e495c8f814c8829986f7a063410