[rn][win] Fix tab behavior when leaving FocusZoneView
Summary:
Our FocusZoneView implementation had a bug on Windows where hitting tab to leave the FocusZoneView would focus the next control that's below the FocusZoneView while the tab order might instead mean that a control to the right of the FocusZoneView should receive focus. To fix this we now rely on Xaml's TabFocusNavigation and XYFocusKeyboardNavigation properties to implement the Windows FocusZoneView.

One example where this issue occurred is if you were focused in the thread list and hit tab it would skip focusing on the chat header buttons.

I had to workaround two Xaml bugs:
1. TabFocusNavigation doesn't propagate the Once value past Xaml ScrollViewers (the native ScrollView on Windows). To workaround this, in ScrollViewManager I check if a parent set TabFocusNavigation(Once) then I re-set it on the ScrollView.Content().
2. When TabFocusNavigation(Once) is set, shift+tab doesn't work correctly when the subviews aren't directly children of the parent, which happens in a lot of places, e.g. FlatList. To workaround this I had to re-implement Shift+tab in a KeyDown handler to fix it. https://github.com/microsoft/microsoft-ui-xaml/issues/1363

This change also means that our Windows FocusZoneView now behaves like a grid, so the focusDirection prop is ignored on Windows. We have a follow-up task T77143805 for the Mac FocusZoneView to support grid layouts.

Test Plan: See Test Plan in FBS diff: D28732436

Reviewers: skyle, ericroz, mylando

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D29044280

Tasks: T89082515