[rn][win] Enable 'inverted' prop on VirtualizedList on Windows
Summary:
When we initially worked on ScrollView inversion for Zeratul chat threads on Windows, we had no way of creating a private fork of the react-native-windows NPM package and thus could not modify VirtualizedList to support our approach for native inversion.

Now that we have tooling to publish react-native-windows to the yarn offline-mirror, this diff modifies VirtualizedList.windows.js to use flexDirection to reverse list content, inverts the frame metrics for rendered items when appropriate, and inverts the scroll metrics that are used to determine when we've reached the end of the list or need to load additional content.

A few additional notes:
1. There is still work to do to re-enable virtualization using this approach
2. Microsoft has recently forked VirtualizedList into their own package for react-native-windows which won't land until RN v0.66

Test Plan:
Fixes issue with scrollToEnd:
|Before|After|
|https://pxl.cl/1N6tp|https://pxl.cl/1N6tn| 

Retains functionality described in:
- https://fb.quip.com/4zzqAVGEOLK1#GcCAAAlNaKJ

Reviewers: skyle, lyahdav, mylando, mowang

Reviewed By: lyahdav

Subscribers: eliwhite

Differential Revision: https://phabricator.intern.facebook.com/D29916475

Signature: 29916475:1629406453:b2459131ea3cd5450bcc44bcdcd00060fb7e5480