Manual cherry-pick react-native-windows#8495, 439b8a54: Adds hook for view managers to capture pointer events
Summary:
Currently, there is no way for ViewManagers to capture the pointer to perform it's own pointer event handling. This change introduces a new interface, `IViewManagerWithPointerEvents`, so third party modules can be notified when the ReactRootView receives pointer events. The behavior is as follows:

`ViewManagerBase` adds a virtual method:
`OnPointerEvent(object sender, ReactPointerEventArgs args)`

We have a `ReactPointerEventArgs` wrapper around `PointerRoutedEventArgs` that view managers can use to:
1. Call `PreventDefault` to suppress the corresponding React event from being emitted
2. Call `ReleaseCapture` to force the TouchEventEmitter to release pointer capture (triggering an `onTouchCancel` event to React Native)
3. Update the `Target` property to refine for hit testing
 
The `ReactPointerEventArgs` wrapper also includes an enum `Kind` value for the type of pointer event so we can reduce the number of methods that need to be implemented for this interface.


Differential Revision: https://phabricator.intern.facebook.com/D30680827