Ensure correct view gets transparent brush
We set a transparent brush on any view that registers for `onMouseEnter`
or `onMouseLeave` to ensure the views are hit testable.

This change ensures that the transparent background is applied to the
border control, if applicable, and not the view panel. I couldn't
reproduce any bugs related to this, but it seemed inconsistent to apply
the `backgroundColor` prop to `ViewPanel::ViewBackground` and this transparent
default to the `Panel::Background` property.