Adds VM interface for custom measurement
While the default self measure function implemented in ViewManagerBase
works in most scenarios, it may not work for all use cases. This change
introduces an ABI VM interface to callback when Yoga measurement is
being performed.

Since Yoga self measure functions are C-style function pointers, we
can't pass instance methods to Yoga. Thus we needed to use a trampoline
approach to provide a static method that can subsequently call the
appropriate measure instance method on the view manager (that has access
to the IViewManagerWithNativeLayout intsance as well). In order to
support this, we needed to add a pointer to the view manager to the
YogaContext that gets passed back to the measure function.

Adding the instance method will add 8 bytes to each Yoga node, which is
not great (especially for apps with many views), but we can probably
reduce this impact by only setting YogaContext nodes on self measuring
view managers, which I will do in a separate diff.