[rn][win][Fabric] Add UseWinUIFabric experimental option to build
Summary:
We need a separate experimental option for Fabric WinUI so we don't
stomp on Microsoft's UseFabric flag for composition Fabric. This change
adds the `UseWinUIFabric` option to ExperimentalFeatures.props and wires
it up in the appropriate .vcxproj and .vcxproj.filters files so the app
will compile.

Test Plan: See test plan in D41662436

Reviewers: lyahdav, shawndempsey, chpurrer, helenistic

Reviewed By: chpurrer

Subscribers: orvid

Differential Revision: https://phabricator.intern.facebook.com/D41662345

Tasks: T138538931