[rn][win][Fabric] Add event data to Touch.h for Desktop
Summary:
In Paper, because the event data was serialized for `receiveTouches`, we
could add arbitrary data to the native event payload. In Fabric, we have
to explicitly wire up these values. This change adds `button` and modifier
keys to the `receiveTouches` event payload.

Test Plan: See test plan in D41748160


Reviewers: lyahdav, shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lyahdav

Differential Revision: https://phabricator.intern.facebook.com/D41748272

Tasks: T138842035, T138841596