[rn][win][Fabric] Fix focusable prop for View
Summary:
The focusable prop was not wired up correctly because ViewProps did not
include the focusable prop on Windows. Now that the focusable prop is
enabled for Windows, this change fixes up the WinUI Fabric
implementation to properly enable and disable focusability.

Test Plan: https://pxl.cl/2mxw9

Reviewers: lyahdav, shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D41841964

Tasks: T138599594