[rn][win][Fabric] Refactor pointer events handling
Summary:
This change isolates the Fabric specific touch and pointer event handling
from the Paper implementation, as this implementation requires an
approach that is mostly similar but differs in key ways in a few places.

In addition to isolating the Fabric touch event handling into its own
class, this change implements proper pointerEvents prop handling, which
also requires customization in the stacking context behavior for
Fabric's view flattening. Specifically, we can only flatten materialized
views if they use pointerEvents == 'auto' on Windows, as Windows uses a
"bottom-up" approach via the "OriginalSource" property on the
RoutedPointerEventArgs object, which already accounts for transforms and
is presumably more performant because we don't need to search for the
target twice.

Test Plan: https://pxl.cl/2njST

Reviewers: lyahdav, shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lyahdav

Differential Revision: https://phabricator.intern.facebook.com/D42107697

Tasks: T138847664, T138852929, T138852966