[rn][win][Fabric] Add selectable prop to Text
Summary:
This wires up a selectable prop implementation largely based on the
Paper implementation for the Text native component. It relies on the
OnPointerEvent API on BaseComponentView to determine if selection has
changed since the pointer was initially pressed, and if not, marks the
event as an "End" event so it triggers a touch event in React.

Test Plan: https://pxl.cl/2nlJN

Reviewers: lyahdav, shawndempsey, chpurrer, lefever, helenistic

Reviewed By: lyahdav

Differential Revision: https://phabricator.intern.facebook.com/D42108420

Tasks: T138599349