[rn][win][Fabric] Adds keyUp/DownEvents implementation
Summary:
In Paper, `keyDownEvents` and `keyUpEvents` are used as a declarative
mechanism to mark certain pointer events as handled. They are the
effective equivalent of `preventDefault` in Web events, but account for
the fact that events are handled asynchronously, so it's generally not
possible to prevent native handling without some kind of declarative
API.

This change implements `keyDownEvents` and `keyUpEvents` in Fabric.

Test Plan: https://pxl.cl/2pm5M

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42178821

Tasks: T140576131