[rn][win][Fabric] Add `onMouseEnter`/`onMouseLeave` events
Summary:
We could force Fabric desktop apps to use `onPointerEnter` and
`onPointerLeave`, but it's just as easy to add `onMouseEnter` and
`onMouseLeave` events for backward compatibility for now.

This adds the `onMouseEnter/Leave` events to the ViewEventEmitter for
now and wires these events up to the touch event handler for Fabric.

Test Plan:
RNTester > Mouse APIs example:
https://pxl.cl/2pm1X


Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42387129

Tasks: T138842072