[rn][win][Fabric] Unsubscribe EventBeat callback
Summary:
The EventBeat implementation for Windows depends on the
xaml::Media::CompositionTarget::Rendering event, which has some heavy
performance implications (namely subscribing it adds idle CPU usage and
prevents the app from entering low power mode).

This change modifies the core AsynchronousEventBeat (arguably, we could
also create our own EventBeat implementation from scratch, but we can
decide that at a later date) to subscribe the Rendering event when a
request occurs and unsubscribe it when a request is processed.

Test Plan:
Notice the idle CPU usage in the before and after plan when the Playground app is minimized:
|Before|After|
| https://pxl.cl/2ptSd | https://pxl.cl/2ptSb |

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42421926

Tasks: T139243050