[rn][win][Fabric] Refactor updateLayoutMetrics for overflow: hidden
Summary:
overflow: 'hidden' will be implemented on all native components. It is
implemented in Paper by calling SetLayoutProps on ViewManagerBase, which
updates the Clip property rect (if needed) to the size of the view as
computed by Yoga. SetLayoutProps is the effective equivalent of
updateLayoutMetrics. As long as we're implementing updateLayoutMetrics
in BaseComponentView, we might as well add the fairly common logic for
setting the Width/Height/Top/Left props of a view there as well.

Test Plan: https://pxl.cl/2pN7J

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42485990

Tasks: T141838989