[rn][win][Fabric] Implement clearTextOnSubmit for single-lined TextInput
Summary:
Implemented clearTextOnSubmit for singled-lined, non-custom TextBoxes. If we have determined that we should submit, we use emitter to call onSubmitEditing (https://www.internalfb.com/diff/D42513564). If we clearTextOnSubmit is passed in, we also clear text. The following still needs to be done:
1) Implement support for multi-line and encrypted textBox
2) Implement support for custom submit keys 

Test Plan:
I altered RNTester to have one textInput with clearTextOnSubmit and one without:
https://pxl.cl/2pTJQ

Reviewers: shawndempsey, ericroz, lefever

Reviewed By: ericroz

Differential Revision: https://phabricator.intern.facebook.com/D42514082

Tasks: T138851514