[rn][win][Fabric] Implements numberOfLines prop on Text
Summary:
It doesn't appear that DWrite's IDWriteTextLayout interface has
functionality for number of lines. So, instead we make two calls to
GetLineMetrics, the first retrieves the number of lines, the second gets
the line metrics for the given line. We return the sum of the line
heights up to the maximum number of lines for the height of the
TextBlock.

Test Plan:
Text example in RNTester:
{F848276591}

Simple example where line heights might differ:
https://pxl.cl/2q4pb

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42549145

Tasks: T138543086