[rn][win][Fabric] Implement textAlignment prop for TextInput
Summary:
Checks if textAlignment prop has changed in inputProps. If it has, we either call updateTextAlignment (if it exists), or we remove it (if the new textAlignment is undefined).

updateTextAlignment follows the same format as Paper's setTextAlignment. We check for if the new alignment is left, right, center, or justified, else we set it to detect from content.

Test Plan:
Playground test depicting setting textAlignment to right, and then to null:
https://pxl.cl/2nFgj

The hacky solution to check between setting a value and removing it:
```
const Bootstrap = () => {
  const [value, onChangeValue] = React.useState('right');

  return (
    <View>
      <Button
        title={'Click me to change textAlignment'}
        onPress={() => {
          value === 'right' ? onChangeValue('left') : onChangeValue('right');
        }}
      />
      <TextInput style={value === 'right' ? {textAlign: value} : {}} />
    </View>
  );
};
```

Reviewers: shawndempsey, ericroz, chpurrer

Reviewed By: ericroz

Differential Revision: https://phabricator.intern.facebook.com/D42216236