[rn][win][Fabric] Implement backgroundColor prop
Summary:
Checks if backgroundColor prop has changed in inputProps. If it has, we either set the field to the new backgroundColor (if it exists), or we set it to null (if the new backgroundColor is undefined).

At the moment, there is a bug on paper's implementation of backgroundColor that doesn't fully remove the color when it is set to undefined. We are replicating it in Fabric and filing a bug with Microsoft. Bug filed: https://github.com/microsoft/react-native-windows/issues/11063

Test Plan:
Testing on playground:

https://pxl.cl/2psK2

Repro this playground instance with:

```
const Bootstrap = () => {
  const [value, onChangeValue] = React.useState('red');

  return (
    <View>
      <Button
        title={'Click me to change backgroundColor'}
        onPress={() => {
          value === 'red' ? onChangeValue('nah') : onChangeValue('red');
        }}
      />
      <TextInput style={value === 'red' ? {backgroundColor: value} : {}} />
    </View>
  );
};

```

Reviewers: shawndempsey, ericroz, chpurrer

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42333352

Tasks: T138851302