[rn][win][Fabric] Avoid text truncation
Summary:
Occasionally, minor differences in the text layout algorithm for XAML
and TextLayoutManager in Fabric lead to text truncation. To minimize the
likelihood of this occurrence, we get the layout metrics for text twice.
The first measurement retrieves the initial bounds for the Text. The
second measurement bounds the width to 1 logical pixel more than the
initial measurement (if possible). We use the "expanded" additional
pixel width as the width for the Text to avoid truncation.

Test Plan:
|Before|After|
| {F849126753} |{F849124622}|

Reviewers: shawndempsey, chpurrer, lefever, helenistic

Reviewed By: chpurrer

Differential Revision: https://phabricator.intern.facebook.com/D42577870

Tasks: T140528185