[rn][win][Fabric] Adds snapTo* and pagingEnabled props
Summary:
I don't believe snapTo* or pagingEnabled props ever worked in Paper, but
just to be on the safe side, this wires up the
ScrollViewUWPImplementation and SnapPointManagingContentControl for
Fabric the same way it's wired up for Paper. This will also make it
easier to integrate our custom native inverted solution that we use in
our fork of react-native-windows for Paper.

Test Plan: See test plan in D42787560.

Reviewers: shawndempsey, lefever, helenistic

Reviewed By: helenistic

Differential Revision: https://phabricator.intern.facebook.com/D42777262

Tasks: T143438685